requirements.txt
python3-netsnmp==1.1a1

